+++ TheatricalPlayersRefactoringKata/Performances/HistoryPerformance.cs
@@ -8,7 +8,7 @@ public class HistoryPerformance : BasePerformance
     public override int Audience { get; set; }
     public override IPlay Play { get; set; }
 
-    public override int CalculateAmmount()
+    public override decimal CalculateAmmount()
     {
         var tragedy = new TragedyPerformance(Play, Audience);
